refactor: split up createFetcher module
