+++ src/libs/createFetchers/modules/createFetcher/index.ts
@@ -1,0 +1,21 @@
+import axios, { AxiosRequestConfig, AxiosResponse } from "axios";
+import {CreateFetcher, RequestConfig} from "./types";
+
+const apiClient = axios.create()
+
+export const createFetcher: CreateFetcher = (host, method, sharedHeaders?) =>
+    <Request, Response, Path>(config: RequestConfig<Request, Path>) => {
+        const {path, body, query, headers} = config
+        return apiClient.request<AxiosRequestConfig<Request>, AxiosResponse<Response>>({
+            method: method,
+            url: `${host}${path}`,
+            data: body,
+            params: query,
+            headers: {
+                "Content-Type": "application/json",
+                "Accept": "application/json",
+                ...sharedHeaders,
+                ...headers,
+            },
+        })
+    }
